+++ test/misc_test.rb
@@ -17,6 +17,11 @@ class MiscTest < Minitest::Test
     assert_unsafe RenameSchema
   end
 
+  def test_rename_enum_value
+    skip unless postgresql?
+    assert_unsafe RenameEnumValue
+  end
+
   def test_create_table_force
     assert_unsafe CreateTableForce
   end